Criado a tela de cadastro de clientes, realizado ajuste na tela de contato, criado a classe ClienteMb, criado Filter e realizada algumas alterações do Util.

+++ embelezzejsf/src/br/senai/sc/ti20131n/pw/embelezzejsf/dao/ProdutoDao.java
@@ -32,7 +32,7 @@ public class ProdutoDao {
 		return entityManager.find(Produto.class, ID);
 	}
 	
-	public void excluirProdutoPorId(Long l){
+	public void excluirProdutoPorId(Integer l){
 		Produto produto = entityManager.getReference(Produto.class, l);
 		entityManager.remove(produto);
 	}
